spin out all the properties since the file was getting huge, dropdown enhancements
